add additional loop to check_matchingNeighbors
original implementation was not checking the final
column properly. it would only check the right
neighbor which results in error and would skip
checking the bottom neighbor.

the original nested for loops check the first 3
columns and every row for matching values.The new
loop will check the last column for matching
values.